[UPDATE] add modules, config, server, utils

+++ utils/parse.go
@@ -5,7 +5,7 @@ import (
 	"strconv"
 )
 
-func ParseStringToInt(s string) int {
+func ParseStringToInt(s string) int64 {
 	result, err := strconv.ParseInt(s, 10, 64)
 	if err != nil {
 		log.Fatal("Error parsing string to int64 failed")
